add product controller - 19 06

+++ src/main/java/com/pawilion/repository/CustomerRepository.java
@@ -3,4 +3,4 @@ package com.pawilion.repository;
 import com.pawilion.model.Customer;
 import org.springframework.data.repository.reactive.ReactiveCrudRepository;
 
-public interface CustomerRepository extends ReactiveCrudRepository<Customer, Integer> {}
+public interface CustomerRepository extends ReactiveCrudRepository<Customer, Long> {}
